cast the types in correctly debug message

+++ libswfdec/swfdec_bits.c
@@ -743,7 +743,7 @@ out:
   } else {
     if (buffer->length < z.total_out) {
       SWFDEC_WARNING ("Not enough data decompressed: %lu instead of %"G_GSIZE_FORMAT" expected",
-	  z.total_out, buffer->length);
+	  (gulong) z.total_out, buffer->length);
       memset (buffer->data + z.total_out, 0, buffer->length - z.total_out);
     }
   }
